Task #5726: Created Cobalt tag 1.6.6

+++ RTCP/Cobalt/GPUProc/test/cuda/tFIR_Filter.cc
@@ -192,7 +192,7 @@ int test()
   hKernel.setArg(1, devSampledData);
   hKernel.setArg(2, devFirWeights);
   hKernel.setArg(3, devHistoryData);
-  unsigned subbandIdx = 0;
+  size_t subbandIdx = 0;
   hKernel.setArg(4, subbandIdx);
 
   // Run the kernel
